+++ libXaw/src/AsciiSink.c
@@ -1653,7 +1653,7 @@ GetGC(AsciiSinkObject sink)
     XGCValues values;
 
     /* XXX We dont want do share a gc that will change the clip-mask */
-    values.clip_x_origin = (long)sink;
+    values.clip_x_origin = (long)(uintptr_t)sink;
     values.clip_mask = None;
     values.font = sink->ascii_sink.font->fid;
     values.graphics_exposures = False;
